Reimplemented bug fix and added tests
