+++ src/components/shared/transactionResult/index.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { withRouter } from 'react-router';
 import { withTranslation } from 'react-i18next';
-import { transactionBroadcasted, resetTransactionResult, secondPassphraseRemoved } from '@actions';
+import { transactionBroadcasted, resetTransactionResult } from '@actions';
 import TransactionResult from './transactionResult';
 
 const mapStateToProps = state => ({
@@ -12,7 +12,6 @@ const mapStateToProps = state => ({
   account: {
     ...state.account.info[state.settings.token.active],
     passphrase: state.account.passphrase,
-    secondPassphrase: state.account.secondPassphrase,
     hwInfo: state.account.hwInfo,
   },
   network: state.network.networks[state.settings.token.active],
@@ -21,7 +20,6 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = {
   transactionBroadcasted,
   resetTransactionResult,
-  secondPassphraseRemoved,
 };
 
 export default compose(
